Move source core tests
